requirements.txt
langchain
openai-whisper 
langchain-groq
sounddevice
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-auth
firebase_admin
soundfile
scipy
torch
TTS
